Add documentation/TODOs to AgeCalculationUtility.
This class has some functions that perform computations regarding the
relative maturity of two pawns, and its behavior is not obvious from
inspection. During the course of understanding it, this change adds
documentation for functions and code blocks to explain what the
computations are doing.

Also, a number of potential TODOs have been marked, based on possible
optimizations that could be done and removing assumptions that have
been made in the code that should not have been.

Also adds a very small amount of error checking, to avoid an NRE.